MNT: Move OnlineSurrogateModel to model file
